Restructuring: seperating sudoku solver and visualizer into different files
