+++ tests/conftest.py
@@ -1,6 +1,6 @@
 # tests/conftest.py
 import pytest
-from src.main import SudokuSolver
+from src.sudoku_solver import SudokuSolver
 
 @pytest.fixture
 def valid_board() -> list[list[int]]:
